Добавлено отображение координаты в командную строку

+++ dev/UserInterface/GameField/GameField.h
@@ -10,7 +10,7 @@ class GameField : public QWidget
 public:
     explicit GameField(QWidget *parent = nullptr);
 
-    void getParameters(int *dx, int *dy, int *sizeBoard, int *rectX, int *rectY, QVector<char> *character, QVector<char> *numbers);
+    void getParameters(float *dx, float *dy, float *sizeBoard, float *rectX, float *rectY, QVector<char> *character, QVector<char> *numbers);
     void paintEvent(QPaintEvent *event) override;
     void mousePressEvent(QMouseEvent *event) override;
